chore: colocate flow utilities
