+++ apps/riven/lib/message-queue/flows/request-content-services/utilities/process-requested-item.ts
@@ -3,10 +3,10 @@ import { ItemRequest } from "@repo/util-plugin-sdk/dto/entities";
 import { ValidationError, validateOrReject } from "class-validator";
 import z from "zod";
 
-import { database } from "../../../database/database.ts";
-import { logger } from "../../../utilities/logger/logger.ts";
+import { database } from "../../../../database/database.ts";
+import { logger } from "../../../../utilities/logger/logger.ts";
 
-import type { MainRunnerMachineIntake } from "../index.ts";
+import type { MainRunnerMachineIntake } from "../../../../state-machines/main-runner/index.ts";
 import type { ContentServiceRequestedResponse } from "@repo/util-plugin-sdk/schemas/events/content-service-requested.event";
 
 export type ProcessRequestedItemInput = {
